move editmission state to redux store
