+++ client/src/components/header/Header.jsx
@@ -7,12 +7,12 @@ import { Button, Box, Typography } from '@material-ui/core';
 import AddIcon from '@material-ui/icons/Add';
 
 // import ExpandLessIcon from '@material-ui/icons/ExpandLess';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 
 
 import { useStyles } from '../../hooks/screenSize';
 
-import { showEditMissionModal } from '../../redux/showEditMissionModal';
+import { showEditMissionModal } from '../../redux/editMission.slice';
 import DisplayOrganizationVision from '../organization_vision/org_display_vision/VisionDisplay';
 
 const Header = () => {
@@ -22,6 +22,8 @@ const Header = () => {
   const [isAdmin] = useState(true);
 
   const classes = useStyles();
+
+  const {missionText} = useSelector(state => state.editMission);
 
   return (
     <Box
@@ -53,7 +55,7 @@ const Header = () => {
           )}
         </Box>
         <Typography className={classes.root} style={{ padding: 12, backgroundColor: 'white' }}>
-          Training A Million Youths Yearly
+          {missionText}
         </Typography>
       </Box>
       <DisplayOrganizationVision />
