+++ client/src/redux/editMission.slice.js
@@ -4,14 +4,18 @@ export const editMissionSlice = createSlice({
   name: 'editMission',
   initialState: {
     showMission: false,
+    missionText: 'Training A Million Youths Yearly'
   },
   reducers: {
     showEditMissionModal: (state) => {
       state.showMission = !state.showMission;
     },
+    editMissionText: (state, action) => {
+      state.missionText = action.payload
+    }
   },
 });
 
-export const { showEditMissionModal } = editMissionSlice.actions;
+export const { showEditMissionModal, editMissionText } = editMissionSlice.actions;
 
 export default editMissionSlice.reducer;
