+++ client/src/redux/store.js
@@ -2,13 +2,13 @@ import { configureStore } from '@reduxjs/toolkit';
 import deleteGoalReducer from './deleteGoal.slice';
 import newGoalReducer from './newGoalSlice';
 import visionReducer from './organizationVision.slice';
-import editMissionReducer from './showEditMissionModal';
+import editMissionReducer from './editMission.slice';
 import createGoalModalReducer from './toggleCreateGoalModal.slice';
 import editGoalModalReducer from './toggleEditGoalModal.slice';
 
 const store = configureStore({
   reducer: {
-    showMission: editMissionReducer,
+    editMission: editMissionReducer,
     organizationVision: visionReducer,
     toggleCreateGoalModal: createGoalModalReducer,
     toggleEditGoalModal: editGoalModalReducer,
